refactor: utils에 넣어두었던 함수들 movie-list로 이동
