--- src/utils/MovieUtil.ts
@@ -1,5 +1,0 @@
-import { IMAGE_URL } from "../constants/constant";
-
-export const createImageUrl = (baseUrl: string, imageUrlPath: string) => {
-  return imageUrlPath.length > 0 ? `${baseUrl}${imageUrlPath}` : IMAGE_URL.DEFAULT_THUMBNAIL_IMAGE_URL;
-};
